requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
httpx>=0.27.0
python-dotenv>=1.0.0

# LangChain / LangGraph / Groq
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.2.0

# Search: DuckDuckGo Lite (BeautifulSoup) + duckduckgo_search fallback
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0

# UI
streamlit>=1.40.0
requests>=2.32.0
